+++ lib/src/features/restrict_apps/method_channel/method_names.dart
@@ -11,8 +11,6 @@ class RestrictionsMethodNames {
   static const String getModesConfig = 'getModesConfig';
   static const String isRestrictionSessionActiveNow =
       'isRestrictionSessionActiveNow';
-  static const String isRestrictionSessionConfigured =
-      'isRestrictionSessionConfigured';
   static const String pauseEnforcement = 'pauseEnforcement';
   static const String resumeEnforcement = 'resumeEnforcement';
   static const String startModeSession = 'startModeSession';